+++ mobile/android/android-components/components/concept/tabstray/src/main/java/mozilla/components/concept/tabstray/TabsTray.kt
@@ -5,7 +5,6 @@
 package mozilla.components.concept.tabstray
 
 import android.view.View
-import mozilla.components.browser.session.Session
 import mozilla.components.support.base.observer.Observable
 
 /**
@@ -19,58 +18,45 @@ interface TabsTray : Observable<TabsTray.Observer> {
         /**
          * A new tab has been selected.
          */
-        fun onTabSelected(session: Session)
+        fun onTabSelected(tab: Tab)
 
         /**
          * A tab has been closed.
          */
-        fun onTabClosed(session: Session)
+        fun onTabClosed(tab: Tab)
     }
 
     /**
-     * Displays the given list of sessions.
+     * Updates the list of tabs.
      *
-     * This method will be invoked with the initial list of sessions that should be displayed.
-     *
-     * @param sessions The list of sessions to be displayed.
-     * @param selectedIndex The list index of the currently selected session.
-     */
-    fun displaySessions(sessions: List<Session>, selectedIndex: Int)
-
-    /**
-     * Updates the list of sessions.
-     *
-     * Calling this method is usually followed by calling onSession*() methods to indicate what
+     * Calling this method is usually followed by calling onTabs*() methods to indicate what
      * exactly has changed. This allows the tabs tray implementation to animate between the old and
      * new state.
-     *
-     * @param sessions The list of sessions to be displayed.
-     * @param selectedIndex The list index of the currently selected session.
      */
-    fun updateSessions(sessions: List<Session>, selectedIndex: Int)
+    fun updateTabs(tabs: Tabs)
 
     /**
-     * Called after updateSessions() when <code>count</code> number of sessions are inserted at the
+     * Called after updateTabs() when <code>count</code> number of tabs are inserted at the
      * given position.
      */
-    fun onSessionsInserted(position: Int, count: Int)
+    fun onTabsInserted(position: Int, count: Int)
 
     /**
-     * Called after updateSessions() when <code>count</code> number of sessions are removed from
+     * Called after updateTabs() when <code>count</code> number of tabs are removed from
      * the given position.
      */
-    fun onSessionsRemoved(position: Int, count: Int)
+    fun onTabsRemoved(position: Int, count: Int)
 
     /**
-     * Called after updateSessions() when a session changes it position.
+     * Called after updateTabs() when a tab changes it position.
      */
-    fun onSessionMoved(fromPosition: Int, toPosition: Int)
+    fun onTabsMoved(fromPosition: Int, toPosition: Int)
 
     /**
-     * Called after updateSessions() when <code>count</code> number of sessions are updated at the
+     * Called after updateTabs() when <code>count</code> number of tabs are updated at the
      * given position.
      */
-    fun onSessionsChanged(position: Int, count: Int)
+    fun onTabsChanged(position: Int, count: Int)
 
     /**
      * Convenience method to cast the implementation of this interface to an Android View object.