+++ mobile/android/android-components/components/feature/tabs/src/test/java/mozilla/components/feature/tabs/tabstray/TabsFeatureTest.kt
@@ -4,14 +4,15 @@
 
 package mozilla.components.feature.tabs.tabstray
 
-import mozilla.components.browser.session.Session
 import mozilla.components.browser.session.SessionManager
+import mozilla.components.browser.state.state.TabSessionState
+import mozilla.components.browser.state.store.BrowserStore
+import mozilla.components.concept.tabstray.Tabs
 import mozilla.components.feature.tabs.TabsUseCases
 import mozilla.components.support.test.mock
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotEquals
 import org.junit.Test
-
 import org.mockito.Mockito.spy
 import org.mockito.Mockito.verify
 
@@ -20,10 +21,11 @@ class TabsFeatureTest {
     @Test
     fun `asserting getters`() {
         val sessionManager = SessionManager(engine = mock())
+        val store = BrowserStore()
         val presenter: TabsTrayPresenter = mock()
         val interactor: TabsTrayInteractor = mock()
         val useCases = TabsUseCases(sessionManager)
-        val tabsFeature = spy(TabsFeature(mock(), sessionManager, useCases, mock()))
+        val tabsFeature = spy(TabsFeature(mock(), store, useCases, mock()))
 
         assertNotEquals(tabsFeature.interactor, interactor)
         assertNotEquals(tabsFeature.presenter, presenter)
@@ -38,10 +40,11 @@ class TabsFeatureTest {
     @Test
     fun start() {
         val sessionManager = SessionManager(engine = mock())
+        val store = BrowserStore()
         val presenter: TabsTrayPresenter = mock()
         val interactor: TabsTrayInteractor = mock()
         val useCases = TabsUseCases(sessionManager)
-        val tabsFeature = spy(TabsFeature(mock(), sessionManager, useCases, mock()))
+        val tabsFeature = spy(TabsFeature(mock(), store, useCases, mock()))
 
         tabsFeature.presenter = presenter
         tabsFeature.interactor = interactor
@@ -55,10 +58,11 @@ class TabsFeatureTest {
     @Test
     fun stop() {
         val sessionManager = SessionManager(engine = mock())
+        val store = BrowserStore()
         val presenter: TabsTrayPresenter = mock()
         val interactor: TabsTrayInteractor = mock()
         val useCases = TabsUseCases(sessionManager)
-        val tabsFeature = spy(TabsFeature(mock(), sessionManager, useCases, mock()))
+        val tabsFeature = spy(TabsFeature(mock(), store, useCases, mock()))
 
         tabsFeature.presenter = presenter
         tabsFeature.interactor = interactor
@@ -72,19 +76,20 @@ class TabsFeatureTest {
     @Test
     fun filterTabs() {
         val sessionManager = SessionManager(engine = mock())
+        val store = BrowserStore()
         val presenter: TabsTrayPresenter = mock()
         val interactor: TabsTrayInteractor = mock()
         val useCases = TabsUseCases(sessionManager)
-        val tabsFeature = spy(TabsFeature(mock(), sessionManager, useCases, mock()))
+        val tabsFeature = spy(TabsFeature(mock(), store, useCases, mock()))
 
         tabsFeature.presenter = presenter
         tabsFeature.interactor = interactor
 
-        val filter: (Session) -> Boolean = { true }
+        val filter: (TabSessionState) -> Boolean = { true }
 
         tabsFeature.filterTabs(filter)
 
-        verify(presenter).sessionsFilter = filter
-        verify(presenter).calculateDiffAndUpdateTabsTray()
+        verify(presenter).tabsFilter = filter
+        verify(presenter).updateTabs(Tabs(emptyList(), -1))
     }
 }