+++ mobile/android/android-components/components/feature/tabs/src/test/java/mozilla/components/feature/tabs/tabstray/TabsTrayPresenterTest.kt
@@ -7,19 +7,28 @@ package mozilla.components.feature.tabs.tabstray
 import android.view.View
 import androidx.lifecycle.LifecycleOwner
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import mozilla.components.browser.session.Session
-import mozilla.components.browser.session.SessionManager
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.TestCoroutineDispatcher
+import kotlinx.coroutines.test.resetMain
+import kotlinx.coroutines.test.setMain
+import mozilla.components.browser.state.action.TabListAction
+import mozilla.components.browser.state.state.BrowserState
+import mozilla.components.browser.state.state.createTab
+import mozilla.components.browser.state.store.BrowserStore
+import mozilla.components.concept.tabstray.Tabs
 import mozilla.components.concept.tabstray.TabsTray
+import mozilla.components.feature.tabs.ext.toTabs
+import mozilla.components.support.test.ext.joinBlocking
 import mozilla.components.support.test.mock
+import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertTrue
+import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.mockito.Mockito.anyInt
-import org.mockito.Mockito.anyList
-import org.mockito.Mockito.never
 import org.mockito.Mockito.spy
 import org.mockito.Mockito.times
 import org.mockito.Mockito.verify
@@ -27,167 +36,213 @@ import org.mockito.Mockito.verifyNoMoreInteractions
 
 @RunWith(AndroidJUnit4::class)
 class TabsTrayPresenterTest {
+    private val testDispatcher = TestCoroutineDispatcher()
 
-    @Test
-    fun `start and stop will register and unregister`() {
-        val sessionManager: SessionManager = mock()
-        val presenter = TabsTrayPresenter(mock(), sessionManager, mock())
+    @Before
+    fun setUp() {
+        Dispatchers.setMain(testDispatcher)
+    }
 
-        verify(sessionManager, never()).register(presenter)
-        verify(sessionManager, never()).unregister(presenter)
-
-        presenter.start()
-        verify(sessionManager).register(presenter)
-        verify(sessionManager, never()).unregister(presenter)
-
-        presenter.stop()
-        verify(sessionManager).unregister(presenter)
+    @After
+    @ExperimentalCoroutinesApi
+    fun tearDown() {
+        Dispatchers.resetMain()
+        testDispatcher.cleanupTestCoroutines()
     }
 
     @Test
     fun `initial set of sessions will be passed to tabs tray`() {
-        val sessionManager = SessionManager(engine = mock())
-
-        sessionManager.add(Session("https://www.mozilla.org"))
-        sessionManager.add(Session("https://getpocket.com"))
+        val store = BrowserStore(
+            BrowserState(
+                tabs = listOf(
+                    createTab("https://www.mozilla.org", id = "a"),
+                    createTab("https://getpocket.com", id = "b")
+                ),
+                selectedTabId = "a"
+            )
+        )
 
         val tabsTray: MockedTabsTray = spy(MockedTabsTray())
-        val presenter = TabsTrayPresenter(tabsTray, sessionManager, mock())
+        val presenter = TabsTrayPresenter(tabsTray, store, mock())
 
         verifyNoMoreInteractions(tabsTray)
 
         presenter.start()
 
-        assertNotNull(tabsTray.displaySessionsList)
-        assertNotNull(tabsTray.displaySelectedIndex)
+        testDispatcher.advanceUntilIdle()
 
-        assertEquals(0, tabsTray.displaySelectedIndex)
-        assertEquals(2, tabsTray.displaySessionsList!!.size)
-        assertEquals("https://www.mozilla.org", tabsTray.displaySessionsList!![0].url)
-        assertEquals("https://getpocket.com", tabsTray.displaySessionsList!![1].url)
+        assertNotNull(tabsTray.updateTabs)
+
+        assertEquals(0, tabsTray.updateTabs!!.selectedIndex)
+        assertEquals(2, tabsTray.updateTabs!!.list.size)
+        assertEquals("https://www.mozilla.org", tabsTray.updateTabs!!.list[0].url)
+        assertEquals("https://getpocket.com", tabsTray.updateTabs!!.list[1].url)
 
         presenter.stop()
     }
 
     @Test
     fun `tab tray will get updated if session gets added`() {
-        val sessionManager = SessionManager(engine = mock())
-
-        sessionManager.add(Session("https://www.mozilla.org"))
-        sessionManager.add(Session("https://getpocket.com"))
+        val store = BrowserStore(
+            BrowserState(
+                tabs = listOf(
+                    createTab("https://www.mozilla.org", id = "a"),
+                    createTab("https://getpocket.com", id = "b")
+                ),
+                selectedTabId = "a"
+            )
+        )
 
         val tabsTray: MockedTabsTray = spy(MockedTabsTray())
-        val presenter = TabsTrayPresenter(tabsTray, sessionManager, mock())
+        val presenter = TabsTrayPresenter(tabsTray, store, mock())
 
         presenter.start()
 
-        assertEquals(2, tabsTray.displaySessionsList!!.size)
+        testDispatcher.advanceUntilIdle()
 
-        sessionManager.add(Session("https://developer.mozilla.org/"))
+        assertEquals(2, tabsTray.updateTabs!!.list.size)
 
-        assertEquals(3, tabsTray.updateSessionsList!!.size)
+        store.dispatch(
+            TabListAction.AddTabAction(
+                createTab("https://developer.mozilla.org/")
+            )
+        ).joinBlocking()
 
-        verify(tabsTray).onSessionsInserted(2, 1)
+        assertEquals(3, tabsTray.updateTabs!!.list.size)
+
+        verify(tabsTray).onTabsInserted(2, 1)
 
         presenter.stop()
     }
 
     @Test
     fun `tabs tray will get updated if session gets removed`() {
-        val sessionManager = SessionManager(engine = mock())
-
-        val firstSession = Session("https://www.mozilla.org")
-        sessionManager.add(firstSession)
-        val secondSession = Session("https://getpocket.com")
-        sessionManager.add(secondSession)
+        val store = BrowserStore(
+            BrowserState(
+                tabs = listOf(
+                    createTab("https://www.mozilla.org", id = "a"),
+                    createTab("https://getpocket.com", id = "b")
+                ),
+                selectedTabId = "a"
+            )
+        )
 
         val tabsTray: MockedTabsTray = spy(MockedTabsTray())
-        val presenter = TabsTrayPresenter(tabsTray, sessionManager, mock())
+        val presenter = TabsTrayPresenter(tabsTray, store, mock())
 
         presenter.start()
 
-        assertEquals(2, tabsTray.displaySessionsList!!.size)
+        testDispatcher.advanceUntilIdle()
 
-        sessionManager.remove(firstSession)
-        assertEquals(1, tabsTray.updateSessionsList!!.size)
-        verify(tabsTray).onSessionsRemoved(0, 1)
+        assertEquals(2, tabsTray.updateTabs!!.list.size)
 
-        sessionManager.remove(secondSession)
-        assertEquals(0, tabsTray.updateSessionsList!!.size)
-        verify(tabsTray, times(2)).onSessionsRemoved(0, 1)
+        store.dispatch(TabListAction.RemoveTabAction("a")).joinBlocking()
+        testDispatcher.advanceUntilIdle()
+
+        assertEquals(1, tabsTray.updateTabs!!.list.size)
+        verify(tabsTray).onTabsRemoved(0, 1)
+
+        store.dispatch(TabListAction.RemoveTabAction("b")).joinBlocking()
+        testDispatcher.advanceUntilIdle()
+
+        assertEquals(0, tabsTray.updateTabs!!.list.size)
+        verify(tabsTray, times(2)).onTabsRemoved(0, 1)
 
         presenter.stop()
     }
 
     @Test
     fun `tabs tray will get updated if all sessions get removed`() {
-        val sessionManager = SessionManager(engine = mock())
-
-        val firstSession = Session("https://www.mozilla.org")
-        sessionManager.add(firstSession)
-        sessionManager.add(Session("https://getpocket.com"))
+        val store = BrowserStore(
+            BrowserState(
+                tabs = listOf(
+                    createTab("https://www.mozilla.org", id = "a"),
+                    createTab("https://getpocket.com", id = "b")
+                ),
+                selectedTabId = "a"
+            )
+        )
 
         val tabsTray: MockedTabsTray = spy(MockedTabsTray())
-        val presenter = TabsTrayPresenter(tabsTray, sessionManager, mock())
+        val presenter = TabsTrayPresenter(tabsTray, store, mock())
 
         presenter.start()
 
-        assertEquals(2, tabsTray.displaySessionsList!!.size)
+        testDispatcher.advanceUntilIdle()
 
-        sessionManager.removeAll()
+        assertEquals(2, tabsTray.updateTabs!!.list.size)
 
-        assertEquals(0, tabsTray.updateSessionsList!!.size)
+        store.dispatch(TabListAction.RemoveAllTabsAction).joinBlocking()
+        testDispatcher.advanceUntilIdle()
 
-        verify(tabsTray).onSessionsRemoved(0, 2)
+        assertEquals(0, tabsTray.updateTabs!!.list.size)
+
+        verify(tabsTray).onTabsRemoved(0, 2)
 
         presenter.stop()
     }
 
     @Test
     fun `tabs tray will get updated if selection changes`() {
-        val sessionManager = SessionManager(engine = mock())
-
-        sessionManager.add(Session("A"))
-        sessionManager.add(Session("B"))
-        sessionManager.add(Session("C"))
-        val sessionToSelect = Session("D").also { sessionManager.add(it) }
-        sessionManager.add(Session("E"))
+        val store = BrowserStore(
+            BrowserState(
+                tabs = listOf(
+                    createTab("https://www.mozilla.org", id = "a"),
+                    createTab("https://getpocket.com", id = "b"),
+                    createTab("https://developer.mozilla.org", id = "c"),
+                    createTab("https://www.firefox.com", id = "d"),
+                    createTab("https://www.google.com", id = "e")
+                ),
+                selectedTabId = "a"
+            )
+        )
 
         val tabsTray: MockedTabsTray = spy(MockedTabsTray())
-        val presenter = TabsTrayPresenter(tabsTray, sessionManager, mock())
+        val presenter = TabsTrayPresenter(tabsTray, store, mock())
 
         presenter.start()
+        testDispatcher.advanceUntilIdle()
 
-        assertEquals(5, tabsTray.displaySessionsList!!.size)
-        assertEquals(0, tabsTray.displaySelectedIndex)
+        assertEquals(5, tabsTray.updateTabs!!.list.size)
+        assertEquals(0, tabsTray.updateTabs!!.selectedIndex)
 
-        sessionManager.select(sessionToSelect)
+        store.dispatch(TabListAction.SelectTabAction("d")).joinBlocking()
+        testDispatcher.advanceUntilIdle()
 
-        assertEquals(3, tabsTray.updateSelectedIndex)
+        println("Selection: " + store.state.selectedTabId)
+        assertEquals(3, tabsTray.updateTabs!!.selectedIndex)
 
-        verify(tabsTray).onSessionsChanged(0, 1)
-        verify(tabsTray).onSessionsChanged(3, 1)
+        verify(tabsTray).onTabsChanged(0, 1)
+        verify(tabsTray).onTabsChanged(3, 1)
     }
 
     @Test
     fun `presenter will close tabs tray when all sessions get removed`() {
-        val sessionManager = SessionManager(engine = mock())
-
-        sessionManager.add(Session("A"))
-        sessionManager.add(Session("B"))
+        val store = BrowserStore(
+            BrowserState(
+                tabs = listOf(
+                    createTab("https://www.mozilla.org", id = "a"),
+                    createTab("https://getpocket.com", id = "b"),
+                    createTab("https://developer.mozilla.org", id = "c"),
+                    createTab("https://www.firefox.com", id = "d"),
+                    createTab("https://www.google.com", id = "e")
+                ),
+                selectedTabId = "a"
+            )
+        )
 
         var closed = false
 
-        val presenter = TabsTrayPresenter(
-            mock(),
-            sessionManager,
-            closeTabsTray = { closed = true })
+        val tabsTray: MockedTabsTray = spy(MockedTabsTray())
+        val presenter = TabsTrayPresenter(tabsTray, store, closeTabsTray = { closed = true })
 
         presenter.start()
+        testDispatcher.advanceUntilIdle()
 
         assertFalse(closed)
 
-        sessionManager.removeAll()
+        store.dispatch(TabListAction.RemoveAllTabsAction).joinBlocking()
+        testDispatcher.advanceUntilIdle()
 
         assertTrue(closed)
 
@@ -196,27 +251,33 @@ class TabsTrayPresenterTest {
 
     @Test
     fun `presenter will close tabs tray when last session gets removed`() {
-        val sessionManager = SessionManager(engine = mock())
-
-        val session1 = Session("A").also { sessionManager.add(it) }
-        val session2 = Session("B").also { sessionManager.add(it) }
+        val store = BrowserStore(
+            BrowserState(
+                tabs = listOf(
+                    createTab("https://www.mozilla.org", id = "a"),
+                    createTab("https://getpocket.com", id = "b")
+                ),
+                selectedTabId = "a"
+            )
+        )
 
         var closed = false
 
-        val presenter = TabsTrayPresenter(
-                mock(),
-                sessionManager,
-                closeTabsTray = { closed = true })
+        val tabsTray: MockedTabsTray = spy(MockedTabsTray())
+        val presenter = TabsTrayPresenter(tabsTray, store, closeTabsTray = { closed = true })
 
         presenter.start()
+        testDispatcher.advanceUntilIdle()
 
         assertFalse(closed)
 
-        sessionManager.remove(session1)
+        store.dispatch(TabListAction.RemoveTabAction("a")).joinBlocking()
+        testDispatcher.advanceUntilIdle()
 
         assertFalse(closed)
 
-        sessionManager.remove(session2)
+        store.dispatch(TabListAction.RemoveTabAction("b")).joinBlocking()
+        testDispatcher.advanceUntilIdle()
 
         assertTrue(closed)
 
@@ -225,60 +286,70 @@ class TabsTrayPresenterTest {
 
     @Test
     fun `presenter calls update and display sessions when calculating diff`() {
-        val sessionManager = SessionManager(engine = mock())
-
-        sessionManager.add(Session("https://www.mozilla.org"))
-        sessionManager.add(Session("https://getpocket.com"))
+        val store = BrowserStore(
+            BrowserState(
+                tabs = listOf(
+                    createTab("https://www.mozilla.org", id = "a"),
+                    createTab("https://getpocket.com", id = "b")
+                ),
+                selectedTabId = "a"
+            )
+        )
 
         val tabsTray: MockedTabsTray = spy(MockedTabsTray())
-        val presenter = TabsTrayPresenter(tabsTray, sessionManager, mock())
+        val presenter = TabsTrayPresenter(tabsTray, store, mock())
 
         presenter.start()
-        presenter.calculateDiffAndUpdateTabsTray()
+        testDispatcher.advanceUntilIdle()
 
-        verify(tabsTray).updateSessions(anyList(), anyInt())
+        val tabs = BrowserState(
+            tabs = listOf(
+                createTab("https://www.firefox.com", id = "c")
+            )
+        ).toTabs()
+
+        presenter.updateTabs(tabs)
+        testDispatcher.advanceUntilIdle()
+
+        verify(tabsTray).updateTabs(tabs)
     }
 
     @Test
     fun `presenter invokes session filtering`() {
-        val sessionManager = SessionManager(engine = mock())
-
-        sessionManager.add(Session("https://www.mozilla.org"))
-        sessionManager.add(Session("https://getpocket.com", private = true))
+        val store = BrowserStore(
+            BrowserState(
+                tabs = listOf(
+                    createTab("https://www.mozilla.org", id = "a"),
+                    createTab("https://getpocket.com", id = "b", private = true)
+                ),
+                selectedTabId = "a"
+            )
+        )
 
         val tabsTray: MockedTabsTray = spy(MockedTabsTray())
-        val presenter = TabsTrayPresenter(tabsTray, sessionManager, mock(), { it.private })
+        val presenter = TabsTrayPresenter(tabsTray, store, mock(), { it.content.private })
 
         presenter.start()
-        presenter.calculateDiffAndUpdateTabsTray()
+        testDispatcher.advanceUntilIdle()
 
-        assertTrue(tabsTray.displaySessionsList?.size == 1)
+        assertTrue(tabsTray.updateTabs?.list?.size == 1)
     }
 }
 
 private class MockedTabsTray : TabsTray {
-    var displaySessionsList: List<Session>? = null
-    var displaySelectedIndex: Int? = null
-    var updateSessionsList: List<Session>? = null
-    var updateSelectedIndex: Int? = null
+    var updateTabs: Tabs? = null
 
-    override fun displaySessions(sessions: List<Session>, selectedIndex: Int) {
-        displaySessionsList = sessions
-        displaySelectedIndex = selectedIndex
+    override fun updateTabs(tabs: Tabs) {
+        updateTabs = tabs
     }
 
-    override fun updateSessions(sessions: List<Session>, selectedIndex: Int) {
-        updateSessionsList = sessions
-        updateSelectedIndex = selectedIndex
-    }
+    override fun onTabsInserted(position: Int, count: Int) {}
 
-    override fun onSessionsInserted(position: Int, count: Int) {}
+    override fun onTabsRemoved(position: Int, count: Int) {}
 
-    override fun onSessionsRemoved(position: Int, count: Int) {}
+    override fun onTabsMoved(fromPosition: Int, toPosition: Int) {}
 
-    override fun onSessionMoved(fromPosition: Int, toPosition: Int) {}
-
-    override fun onSessionsChanged(position: Int, count: Int) {}
+    override fun onTabsChanged(position: Int, count: Int) {}
 
     override fun register(observer: TabsTray.Observer) {}